added checkout page with typescript

+++ src/components/Cart/Cart.tsx
@@ -1,52 +1,61 @@
-import { useRouter } from "next/router";
-import React from "react";
-import { useEffect } from "react";
-import styled from "styled-components";
+import { useRouter } from 'next/router'
+import React from 'react'
+import { useEffect } from 'react'
+import styled from 'styled-components'
 
-import { devices } from "../../data";
-import { UpdateProducts } from "../../apiCalls/apiCalls";
-import { CartProduct } from "./CartProduct";
-import { EmptyCart } from "./EmptyCart";
+import { devices } from '../../data'
+import { UpdateProducts } from '../../apiCalls/apiCalls'
+import { CartProduct } from './CartProduct'
+import { EmptyCart } from './EmptyCart'
+import {
+  addQuantity,
+  removeProduct,
+  removeQuantity,
+} from '../../redux/slices/cart'
+import { CartProductType, handleRemoveProductType } from './Cart.model'
+import { CartInitialState } from '../../redux/slices/slice.model'
+import { AppDispatch, useAppSelector } from '../../redux/store/store'
+import { handleQuantityType } from './Cart.model'
 
-export const Cart : React.FC = () => {
-
+export const Cart: React.FC = () => {
   const router = useRouter()
 
-  /* const cart = useSelector((state) => state.cart);
-  const { isFetching } = useSelector((state) => state.user);
-  const dispatch = useDispatch();
-  const navigate = useNavigate(); */
+  const cart: CartInitialState = useAppSelector((state) => state.cart)
+  const dispatch = AppDispatch()
 
   /* функция изменения количества продуктов */
-  const handlequantity = (index, changer) => {
-    /* if (changer === "rem" && cart.products[index].quantity >= 2) {
-      dispatch(removeQuantity(index));
-    } else if (changer === "add" && cart.products[index].quantity >= 1) {
-      dispatch(addQuantity(index));
-    } */
-  };
+  const handlequantity: handleQuantityType = (index, changer) => {
+    if (changer === 'rem' && cart.products[index].quantity >= 2) {
+      dispatch(removeQuantity(index))
+    } else if (changer === 'add' && cart.products[index].quantity >= 1) {
+      dispatch(addQuantity(index))
+    }
+  }
 
   /* функция удаления товара */
-  const handleRemoveProduct = (item) => {
-    /* dispatch(removeProduct(item)); */
-  };
+  const handleRemoveProduct: handleRemoveProductType = (
+    index: number,
+    item: CartProductType
+  ) => {
+    dispatch(removeProduct({ index, item }))
+  }
 
   /* отправит пользователя на страницу оформления заказа по клику, если в корзине есть товары */
   const handleCheckout = () => {
-    /* if (cart.quantity > 0) {
-      router.replace("/checkout");
-    } */
-  };
+    if (cart.quantity > 0) {
+      router.replace('/checkout')
+    }
+  }
 
   /* функция для получения всех продуктов и обновления цен и изображений продуктов в корзине */
   useEffect(() => {
-    /* UpdateProducts(dispatch, cart.products); */
-  }, []);
+    UpdateProducts(cart.products, dispatch)
+  }, [])
 
   return (
     <Container>
       {/* если у пользователя нет товаров в корзине, покажет компонент пустой корзины */}
-      {1 === 0 ? (
+      {cart.products.length === 0 ? (
         <EmptyCart />
       ) : (
         <Wrapper>
@@ -56,30 +65,28 @@ export const Cart : React.FC = () => {
             <ProductsTotal>ОБЩИЙ</ProductsTotal>
           </Top>
           <Info>
-            {/* {cart.products?.map((item, index) => (
+            {cart.products?.map((item: CartProductType, index: number) => (
               <CartProduct
-                key={item.id}
+                key={index}
                 item={item}
                 index={index}
                 handlequantity={handlequantity}
                 handleRemoveProduct={handleRemoveProduct}
               />
-            ))} */}
+            ))}
           </Info>
           <Bottom>
-            {/* <Total>₽{cart.total}</Total> */}
+            <Total>₽{cart.total}</Total>
             <Shipping>
               Доставка и налоги рассчитываются при оформлении заказа
             </Shipping>
-            <Button onClick={handleCheckout} >
-              Перейти к оформлению
-            </Button>
+            <Button onClick={handleCheckout}>Перейти к оформлению</Button>
           </Bottom>
         </Wrapper>
       )}
     </Container>
-  );
-};
+  )
+}
 
 const Wrapper = styled.div`
   display: flex;
@@ -88,7 +95,7 @@ const Wrapper = styled.div`
   width: 60%;
   gap: 2em;
   font-size: 16px;
-`;
+`
 
 const Top = styled.div`
   display: flex;
@@ -96,14 +103,14 @@ const Top = styled.div`
   width: 100%;
   border-bottom: 1px solid #dcdcdc;
   color: #9d9d9d;
-`;
+`
 
 const ProductsTitle = styled.h1`
   font-size: 12px;
   font-weight: 700;
   letter-spacing: 2.4px;
   flex: 2;
-`;
+`
 
 const ProductsQauntity = styled.span`
   font-size: 12px;
@@ -111,7 +118,7 @@ const ProductsQauntity = styled.span`
   letter-spacing: 2.4px;
   flex: 1;
   text-align: center;
-`;
+`
 
 const ProductsTotal = styled.span`
   font-size: 12px;
@@ -119,12 +126,12 @@ const ProductsTotal = styled.span`
   letter-spacing: 2.4px;
   flex: 1;
   text-align: right;
-`;
+`
 const Info = styled.div`
   display: flex;
   flex-direction: column;
   gap: 3em;
-`;
+`
 
 const Bottom = styled.div`
   display: flex;
@@ -134,15 +141,15 @@ const Bottom = styled.div`
   padding-top: 1em;
   flex-direction: column;
   border-top: 1px solid #dcdcdc;
-`;
+`
 
 const Total = styled.span`
   font-size: 1em;
-`;
+`
 
 const Shipping = styled.span`
   font-size: 1em;
-`;
+`
 
 const Button = styled.button`
   font-size: 0.9em;
@@ -165,7 +172,7 @@ const Button = styled.button`
     cursor: not-allowed;
     background-color: rgb(244, 112, 38);
   }
-`;
+`
 
 const Container = styled.div`
   display: flex;
@@ -190,4 +197,4 @@ const Container = styled.div`
       text-align: right;
     }
   }
-`;
+`
